Add function for add new books
