+++ src/JS/app.js
@@ -1,4 +1,4 @@
-import { content } from './dom.js';
+import { content, showBooks } from './dom.js';
 
 content();
 
@@ -9,7 +9,6 @@ const ftitle = document.querySelector('.f-title');
 const fauthor = document.querySelector('.f-author');
 const fpages = document.querySelector('.f-pages');
 const fedition = document.querySelector('.f-edition');
-const fread = document.querySelector('.f-read');
 const fsubmit = document.querySelector('.f-submit');
 const bookCard = document.querySelector('.book-card');
 
@@ -18,11 +17,15 @@ function Book(title, author, pages, edition, read) {
   this.author = author;
   this.pages = pages;
   this.edition = edition;
-  this.read = !read;
+  this.read = read;
 }
 
 addBt.addEventListener('click', () => {
-  form.className = 'show'
+  if (form.className === 'init') {
+    form.className = 'show'
+  } else if (form.className === 'show') {
+    form.className = 'init';
+  }
 })
 
 
@@ -30,82 +33,53 @@ const updateLocalStorage = () => {
   localStorage.setItem('myLibrary', JSON.stringify(myLibrary));
 }
 
-fsubmit.addEventListener('click', () => {
+const addNewBook = () => {
+  
   if (ftitle.value === '' || fauthor.value === '' || fpages.value === '' || fedition.value === '') {
     alert('There are data missing, please check it.');
   } else {
-    const book = new Book(ftitle.value, fauthor.value, fpages.value, fedition.value, fread);
+    const book = new Book(ftitle.value, fauthor.value, fpages.value, fedition.value);
     myLibrary.push(book);
     updateLocalStorage();
   }
+  
+}
+
+fsubmit.addEventListener('click', (e) => {
+  e.preventDefault();
+  addNewBook();
+  
 });
 
+// const getLocalStorage = () => {
+//   myLibrary = JSON.parse(localStorage.getItem('myLibrary'));
 
-const readContent = (s) => {
-  let status;
-  if (myLibrary[s].read === true) {
-    status = 'Read';
-  } else {
-    status = 'Not Read';
-  }
-  return status;
-}
-
-const showBooks = () => {
-  bookCard.innerHTML = '';
-  for (let i = 0; i < myLibrary.length; i += 1) {
-    const m = readContent(i);
-    const bookDiv = document.createElement('div');
-    bookDiv.className = 'book-item';
-    const cardTitle = document.createElement('h2');
-    cardTitle.setAttribute('class', 'card-title');
-    cardTitle.innerText = myLibrary[i].title;
-
-    const cardAuthor = document.createElement('h3');
-    cardAuthor.setAttribute('class', 'card-author');
-    cardAuthor.innerText = `Author: ${myLibrary[i].author}`;
-
-    const cardPage = document.createElement('p');
-    cardPage.setAttribute('class', 'card-page');
-    cardPage.innerText = `Number of pages: ${myLibrary[i].pages}`;
-
-    const cardEdition = document.createElement('p');
-    cardEdition.setAttribute('class', 'card-edition');
-    cardEdition.innerText = `Edition: ${myLibrary[i].edition}`;
-
-    const cardRead = document.createElement('p');
-    cardRead.setAttribute('class', 'card-read');
-    cardRead.innerHTML = `${m}`;
-
-    const readButton = document.createElement('button');
-    readButton.className = 'read-btn';
-    readButton.innerHTML = 'Change Status';
-    readButton.setAttribute('onclick', `readStatus(${i})`);
-
-    const deleteBtn = document.createElement('button');
-    deleteBtn.setAttribute('type', 'button');
-    deleteBtn.className = 'delete-btn';
-    deleteBtn.textContent = 'Delete';
-    deleteBtn.setAttribute('onclick', `deleteBook(${i})`);
-
-    bookDiv.appendChild(cardTitle);
-    bookDiv.appendChild(cardAuthor);
-    bookDiv.appendChild(cardPage);
-    bookDiv.appendChild(cardEdition);
-    bookDiv.appendChild(cardRead);
-    bookDiv.appendChild(readButton);
-    bookDiv.appendChild(deleteBtn);
-    bookCard.appendChild(bookDiv);
-  }
-}
-
-const getLocalStorage = () => {
+//   if (myLibrary === null) {
+//     myLibrary = [];
+//   }
+//   for(let i = 0; myLibrary.length; i += 1){
+//     showBooks();
+    
+//   }
+// }
+const render = () => {
   myLibrary = JSON.parse(localStorage.getItem('myLibrary'));
-
-  if (myLibrary === null) {
+  // myLibrary.forEach((book) => {
+  //   console.log(book.title);
+  //   if (myLibrary === null) {
+  //     myLibrary = [];
+  //   }
+  //   showBooks(book);
+  // })
+  if (myLibrary != null){
+ 
+    bookCard.innerHTML = '';
+    for(let i = 0; i < myLibrary.length; i += 1){
+      showBooks(myLibrary[i]);
+    }
+  } else {
     myLibrary = [];
   }
-  showBooks();
 }
 
 const deleteBook = (book) => {
@@ -122,4 +96,10 @@ const readStatus = (s) => {
   getLocalStorage();
 }
 
-getLocalStorage();
+window.onload = () => {
+  render();
+}
+//getLocalStorage();
+
+
+
